@@ -1,25 +1,32 @@
 {
   "name": "yad3w",
   "version": "1.0.0",
-  "description": "React Chart",
+  "description": "D3 React wrapper",
   "main": "js/app.js",
   "dependencies": {
-    "d3-arrays": "0.0.4",
-    "d3-scale": "^0.1.1",
-    "d3-selection": "^0.4.7",
-    "react": "^0.13.3",
-    "react-motion": "^0.1.0"
+    "d3-array": "^1.0.0",
+    "d3-scale": "^1.0.0",
+    "d3-selection": "^1.0.0",
+    "d3-shape": "^1.0.0",
+    "react": "^15.2.0",
+    "react-dom": "^15.2.0",
+    "react-motion": "^0.4.4"
   },
   "devDependencies": {
-    "babel-eslint": "^3.1.23",
-    "babelify": "^6.1.2",
-    "browserify": "^10.2.4",
-    "eslint": "^0.24.0",
-    "eslint-plugin-react": "^2.6.4"
+    "babel-eslint": "^6.1.0",
+    "babel-preset-es2015": "^6.9.0",
+    "babel-preset-react": "^6.11.1",
+    "babel-preset-stage-0": "^6.5.0",
+    "babelify": "^7.3.0",
+    "browserify": "^13.0.1",
+    "standard": "^7.1.2"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Mirco Zeiss <mirco.zeiss@gmail.com>",
-  "license": "MIT"
+  "license": "MIT",
+  "standard": {
+    "parser": "babel-eslint"
+  }
 }
