@@ -1,58 +1,27 @@
-'use strict';
 
-var React = require('react');
+import React from 'react'
 
-
-
-/**
- * CSS
- */
-var style = {
+const style = {
   shapeRendering: 'crispEdges'
-};
+}
 
+export default class YGrid extends React.Component {
 
-
-/**
- * YGrid component
- */
-class YGrid extends React.Component {
-
-
-
-  /**
-   * Default properties
-   */
   static defaultProps = {
     ticks: 10,
     fill: 'none',
     stroke: 'lightgrey',
     strokeWidth: '1'
-  };
-
-
-
-  /**
-   * Constructor function
-   */
-  constructor(props) {
-    super(props);
   }
 
-
-
-  /**
-   * Render component
-   */
-  render() {
-
-    var {height, xScale, ticks} = this.props;
+  render () {
+    var {height, xScale, ticks} = this.props
 
     var lines = xScale.ticks(ticks).map((d, i) =>
       <line
         key={i}
         x1={xScale(d)}
-        y1="0"
+        y1='0'
         x2={xScale(d)}
         y2={height}
         fill={this.props.fill}
@@ -60,21 +29,12 @@ class YGrid extends React.Component {
         strokeWidth={this.props.strokeWidth}
         style={style}
       />
-    );
+    )
 
     return (
       <g>
         {lines}
       </g>
-    );
-
+    )
   }
-
 }
-
-
-
-/**
- * Export component
- */
-module.exports = YGrid;
